# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Road Damage Reporter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Report {
  id: number;
  /** pothole, crack, subsidence, flooding, debris, other */
  damageType: string;
  /** low, medium, high, critical */
  severity: string;
  /** @nullable */
  description?: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  photoBase64?: string | null;
  /** pending, in_review, resolved */
  status: string;
  createdAt: string;
}

export interface ReportInput {
  damageType: string;
  severity: string;
  description?: string;
  latitude: number;
  longitude: number;
  address?: string;
  photoBase64?: string;
}

export interface CountByKey {
  key: string;
  count: number;
}

export interface ReportStats {
  total: number;
  /** Reports submitted in the last 7 days */
  recentCount: number;
  byType: CountByKey[];
  bySeverity: CountByKey[];
  byStatus: CountByKey[];
}

export type ListReportsParams = {
damageType?: string;
severity?: string;
limit?: number;
offset?: number;
};