# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Road Damage Reporter API
 * OpenAPI spec version: 0.1.0
 */

export interface Report {
  id: number;
  /** pothole, crack, subsidence, flooding, debris, other */
  damageType: string;
  /** low, medium, high, critical */
  severity: string;
  /** @nullable */
  description?: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  photoBase64?: string | null;
  /** pending, in_review, resolved */
  status: string;
  createdAt: Date;
}