# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Road Damage Reporter API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  damageType: string;
  severity: string;
  description?: string;
  latitude: number;
  longitude: number;
  address?: string;
  photoBase64?: string;
}