# lib/api-zod/src/generated/types/reportStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Road Damage Reporter API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByKey } from './countByKey';

export interface ReportStats {
  total: number;
  /** Reports submitted in the last 7 days */
  recentCount: number;
  byType: CountByKey[];
  bySeverity: CountByKey[];
  byStatus: CountByKey[];
}